Make cgi more or less like brontasaur.us, and fix fetching
For whatever stupid reason, when I ran the scraper from DreamHost
running Python 2.7, when it came to pulling down the OCM page, the page
that OCM would serve up would have a nice error message in it
complaining that I'm using an older browser and I need to update in
order to experience the joys of modern browsing. Maybe bare 2.7.3 would
work, I never bothered to check because I just added requests to
requirements.pip and I know requests.get seems to work fine. So there.